Improving code quality and adding the time viewing
Merge three calsses to be one and add the date of month into the view
